Move parsing context out of the sd_markdown object for thread safety reasons
If multiple threads call render on the same sd_markdown object at the same time,
there would be contention on the work buffers causing invalid memory states.
This moves any data that is modified during a render from the shared sd_markdown object
to a workdata object that's created per call the render.